requirements.txt
gradio>=4.0.0
groq 
python-docx>=0.8.11
pandas>=1.3.0
numpy>=1.20.0
PyPDF2>=2.0.0
openpyxl>=3.0.9
python-dotenv>=0.19.0
Pillow>=9.0.0
requests>=2.27.1
tiktoken>=0.3.0
